logged in users

+++ controllers/classes/UserManager.php
@@ -12,30 +12,51 @@ class UserManager {
     $login = $data["login"];
     $passwd = $data["passwd"];
  
-    $userId = $db->selectUser($login, $passwd, "users");
+    $user = $db->selectUser($login, $passwd, "users");
+     $userId=$user['id'];
+     $ipAddress = $_SERVER['REMOTE_ADDR'];
+     $logs= $db->getLog($userId);
 
-        if ($userId >= 0)
+     $blockLogin= $db->getBlockLogin($ipAddress);
+
+    //pierwsze logowanie
+     if($blockLogin === []) {
+         $sql = "INSERT INTO block_login VALUES (null, 0, null, null, null,'$ipAddress')";
+         $db->insert($sql);
+     }
+
+     if ($user['access'] === true)
         { //Poprawne dane
 
              session_start();
              $_SESSION['login'] = $login;
              $_SESSION['passwd'] = $passwd;
-             $_SESSION['userId'] = $userId;
-
-
-
-         //usuń wszystkie wpisy historyczne dla użytkownika o $userId
-             $sql="DELETE FROM logged_in_users WHERE userId='$userId'";
-             $db->delete($sql);
+             $_SESSION['userId'] = $user['id'];
 
              $date = $this->getCurrentDate();
-             $id_sesji = session_id();
+             $idSession = session_id();
 
-             $sql="INSERT INTO logged_in_users VALUES ('$id_sesji','$userId','$date')";
+             $sql="INSERT INTO logged_in_users VALUES ('$idSession','$userId','$date',true,null,'$ipAddress')";
              $db->insert($sql);
+        } else {
+            $date = $this->getCurrentDate();
+            $sql = "INSERT INTO logged_in_users (userId, lastUpdate, logSuccess, ipAddress) VALUES ('$userId','$date',false,'$ipAddress')";
+            $db->insert($sql);
+
+            /*$blockLogin = $blockLogin[0];
+            $blockLogin->lastBadLoginNum = $this->getCurrentDate();
+
+            $blockLogin->badLoginNum++;
+            if($blockLogin->badLoginNum > 2){
+            if($date > $blockLogin->tempLock && $blockLogin->tempLock !== NULL);
+            $time= pow($blockLogin->badLoginNum,2)*60;
+            }*/
         }
-    
-    return $userId;
+     if(count($logs) > 9){
+         $sql="DELETE FROM logged_in_users WHERE userId='$userId' AND id = (SELECT MIN(id) FROM logged_in_users)";
+         $db->delete($sql);
+     }
+     return $user;
  }
  
     
@@ -49,9 +70,7 @@ class UserManager {
          setcookie(session_name(),'', time() - 42000, '/');
      }
    session_destroy();
-     
-     $db->delete("DELETE FROM logged_in_users WHERE sessionId='$idSession' ");
-        
+
     header("location: ../index.php");
  }
     function validateLogin($input)
